Added option -deck-level-symmetry
Added a new option -deck-level-symmetry that adds an additional
symmetry breaking in for the free play type. The idea is that when the
play-type is free, we can arbitrarily choose that all cards that end
up on the first level are also first in the deck, and so on. This was
suggested by Ciaran McCreesh at CP 2019.

Unfortunately, this did not help as much as expected and is therefore
not enabled by default.